fix(suite): rewrite settings DAC modal to actually use the thunk result
